backend/requirements/base.txt
Django==4.0.3
django-ninja==0.17.0
python-environ==0.4.54
psycopg2-binary==2.8.6
whitenoise==5.2.0
django-cors-headers==3.11.0